# gateway/src/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Tools API Gateway
  description: |
    Unified API Gateway for the MCP Tools ecosystem providing access to Kanban boards,
    Wiki pages, Memory graphs, Web scraping, and Analytics services.
    
    This gateway aggregates multiple MCP (Model Context Protocol) servers into a single
    REST API with WebSocket support for real-time collaboration.
  version: 1.0.0
  contact:
    name: MCP Tools Support
    url: https://github.com/tylercoles/mcp-tools
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:6100
    description: Development server
  - url: http://gateway:6100
    description: Docker container server
paths:
...
  /api/v1/search:
    post:
      summary: Unified search across all content sources
      description: |
        Performs comprehensive search across Memory thoughts, Kanban cards, Wiki pages,
        and scraped content with advanced filtering, semantic search, and result ranking.
      tags:
        - Search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnifiedSearchRequest'
            example:
              query: "machine learning algorithms"
              filters:
                content_types: ["wiki_page", "scraped_page"]
                date_from: "2024-01-01T00:00:00Z"
                min_quality: 0.7
              sort: "relevance"
              pagination:
                page: 1
                limit: 20
              use_semantic: true
              include_preview: true
      responses:
        '200':
          description: Search results successfully retrieved
          headers:
            X-Search-Processing-Time:
              description: Processing time in milliseconds
              schema:
                type: integer
            X-Search-Results-Count:
              description: Total number of results found
              schema:
                type: integer
            X-Search-Cache-Hit:
              description: Whether results were served from cache
              schema:
                type: boolean
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedSearchResponse'
        '400':
          description: Invalid search request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchError'
        '408':
          description: Search request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchError'
        '500':
          description: Search service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchError'
        '503':
          description: Search service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchError'
  /api/v1/search/suggestions:
    get:
      summary: Get search query suggestions
      description: |
        Provides intelligent search suggestions based on query prefix,
        popular searches, and personalized recommendations.
      tags:
        - Search
      parameters:
        - name: q
          in: query
          description: Query prefix for suggestions
          schema:
            type: string
            maxLength: 100
          example: "machine"
        - name: limit
          in: query
          description: Maximum number of suggestions to return
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
        - name: user_id
          in: query
          description: User ID for personalized suggestions
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Search suggestions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchSuggestionsResponse'
        '400':
          description: Invalid suggestions request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchError'
        '503':
          description: Suggestions service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchError'
  /api/v1/search/analytics:
    get:
      summary: Get search analytics data
      description: |
        Returns comprehensive search analytics including performance metrics,
        popular queries, and user search patterns.
      tags:
        - Search
      parameters:
        - name: user_id
          in: query
          description: Filter analytics by user ID
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          description: Start date for analytics period
          schema:
            type: string
            format: date-time
        - name: date_to
          in: query
          description: End date for analytics period
          schema:
            type: string
            format: date-time
        - name: include_performance
          in: query
          description: Include performance metrics in response
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Analytics data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAnalyticsResponse'
        '400':
          description: Invalid analytics request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchError'
        '503':
          description: Analytics service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchError'
  /api/v1/search/cache:
    delete:
      summary: Clear search result cache
      description: |
        Clears the search result cache. This operation should be used sparingly
        and is typically restricted to administrators.
      tags:
        - Search
      responses:
        '200':
          description: Cache cleared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Search cache cleared successfully"
                      cleared_at:
                        type: string
                        format: date-time
        '429':
          description: Too many cache operations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchError'
        '500':
          description: Failed to clear cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchError'
...
components:
  schemas:
...
    UnifiedSearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 1000
          description: Search query text
          example: "machine learning algorithms"
        filters:
          type: object
          properties:
            content_types:
              type: array
              items:
                type: string
                enum: ["scraped_page", "scraped_content_chunk", "wiki_page", "kanban_card", "memory_thought", "code_file", "code_chunk"]
              description: Filter by content types
            date_from:
              type: string
              format: date-time
              description: Filter by date range start
            date_to:
              type: string
              format: date-time
              description: Filter by date range end
            created_by:
              type: string
              format: uuid
              description: Filter by creator user ID
            tags:
              type: array
              items:
                type: string
              maxItems: 10
              description: Filter by tags
            min_quality:
              type: number
              minimum: 0
              maximum: 1
              description: Minimum quality score
            language:
              type: string
              maxLength: 20
              description: Programming language filter
            repository:
              type: string
              maxLength: 100
              description: Repository name filter
        sort:
          type: string
          enum: ["relevance", "date_desc", "date_asc", "title_asc", "title_desc", "quality_desc"]
          default: "relevance"
          description: Sort order for results
        pagination:
          type: object
          properties:
            page:
              type: integer
              minimum: 1
              maximum: 1000
              default: 1
              description: Page number (1-based)
            limit:
              type: integer
              minimum: 1
              maximum: 100
              default: 20
              description: Results per page
            offset:
              type: integer
              minimum: 0
              description: Offset for cursor-based pagination
        use_semantic:
          type: boolean
          default: true
          description: Enable semantic search using embeddings
        use_fuzzy:
          type: boolean
          default: true
          description: Enable fuzzy text matching
        include_preview:
          type: boolean
          default: true
          description: Include content preview in results
        include_highlights:
          type: boolean
          default: true
          description: Include highlighted matches
        user_id:
          type: string
          format: uuid
          description: User ID for personalization
    UnifiedSearchResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            results:
              type: array
              items:
                $ref: '#/components/schemas/SearchResult'
            total_count:
              type: integer
              minimum: 0
              description: Total number of matching results
            pagination:
              type: object
              properties:
                current_page:
                  type: integer
                  minimum: 1
                per_page:
                  type: integer
                  minimum: 1
                total_pages:
                  type: integer
                  minimum: 0
                has_next:
                  type: boolean
                has_prev:
                  type: boolean
            aggregations:
              $ref: '#/components/schemas/SearchAggregations'
            performance:
              type: object
              properties:
                processing_time_ms:
                  type: integer
                  minimum: 0
                vector_search_ms:
                  type: integer
                  minimum: 0
                text_search_ms:
                  type: integer
                  minimum: 0
                documents_searched:
                  type: integer
                  minimum: 0
                gateway_processing_time_ms:
                  type: integer
                  minimum: 0
            suggestions:
              type: array
              items:
                type: object
                properties:
                  query:
                    type: string
                  type:
                    type: string
                    enum: ["spelling", "completion", "related"]
                  confidence:
                    type: number
                    minimum: 0
                    maximum: 1
    SearchResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique result identifier
        type:
          type: string
          enum: ["scraped_page", "scraped_content_chunk", "wiki_page", "kanban_card", "memory_thought", "code_file", "code_chunk"]
          description: Content type
        title:
          type: string
          description: Result title
        preview:
          type: object
          properties:
            text:
              type: string
              description: Preview text
            highlights:
              type: array
              items:
                type: object
                properties:
                  start:
                    type: integer
                  end:
                    type: integer
                  match:
                    type: string
            length:
              type: integer
              description: Preview length in characters
            truncated:
              type: boolean
              description: Whether content was truncated
        url:
          type: string
          description: Direct URL/link to content
        score:
          type: object
          properties:
            relevance:
              type: number
              minimum: 0
              maximum: 1
            semantic_similarity:
              type: number
              minimum: 0
              maximum: 1
            text_match:
              type: number
              minimum: 0
              maximum: 1
            recency_boost:
              type: number
              minimum: 0
              maximum: 1
            quality_score:
              type: number
              minimum: 0
              maximum: 1
        metadata:
          type: object
          properties:
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            created_by:
              type: string
              format: uuid
            tags:
              type: array
              items:
                type: string
            language:
              type: string
            repository:
              type: string
            file_path:
              type: string
            line_range:
              type: object
              properties:
                start:
                  type: integer
                end:
                  type: integer
    SearchAggregations:
      type: object
      properties:
        by_type:
          type: object
          additionalProperties:
            type: integer
        by_date:
          type: object
          properties:
            last_day:
              type: integer
            last_week:
              type: integer
            last_month:
              type: integer
            older:
              type: integer
        top_tags:
          type: array
          items:
            type: object
            properties:
              tag:
                type: string
              count:
                type: integer
        languages:
          type: array
          items:
            type: object
            properties:
              language:
                type: string
              count:
                type: integer
        repositories:
          type: array
          items:
            type: object
            properties:
              repository:
                type: string
              count:
                type: integer
    SearchSuggestionsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            suggestions:
              type: array
              items:
                type: object
                properties:
                  query:
                    type: string
                  type:
                    type: string
                    enum: ["completion", "popular", "related", "corrected"]
                  confidence:
                    type: number
                    minimum: 0
                    maximum: 1
                  context:
                    type: string
                  estimated_results:
                    type: integer
                    minimum: 0
            query_prefix:
              type: string
            total_suggestions:
              type: integer
            personalized:
              type: boolean
            categories:
              type: object
              properties:
                completions:
                  type: integer
                popular:
                  type: integer
                related:
                  type: integer
                corrections:
                  type: integer
    SearchAnalyticsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            timeline:
              type: array
              items:
                type: object
                properties:
                  timestamp:
                    type: string
                    format: date-time
                  search_count:
                    type: integer
                  unique_users:
                    type: integer
                  avg_response_time:
                    type: number
                  top_query:
                    type: string
            popular_queries:
              type: array
              items:
                type: object
                properties:
                  query:
                    type: string
                  count:
                    type: integer
                  avg_results:
                    type: number
                  avg_response_time:
                    type: number
                  percentage:
                    type: number
                  trend:
                    type: string
                    enum: ["up", "down", "stable"]
            summary:
              type: object
              properties:
                total_searches:
                  type: integer
                unique_users:
                  type: integer
                avg_response_time:
                  type: number
                success_rate:
                  type: number
                most_searched_type:
                  type: string
            performance:
              type: object
              properties:
                cache_hit_rate:
                  type: number
                p95_response_time:
                  type: number
                error_rate:
                  type: number
                service_availability:
                  type: object
                  additionalProperties:
                    type: number
            date_range:
              type: object
              properties:
                from:
                  type: string
                to:
                  type: string
                period:
                  type: string
...
    SearchError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              enum: [
                "VALIDATION_ERROR",
                "SEARCH_VALIDATION_ERROR", 
                "SUGGESTIONS_VALIDATION_ERROR",
                "ANALYTICS_VALIDATION_ERROR",
                "SEARCH_RATE_LIMITED",
                "CACHE_RATE_LIMITED",
                "ANALYTICS_RATE_LIMITED",
                "SERVICE_UNAVAILABLE",
                "SEARCH_FAILED",
                "SEARCH_TIMEOUT",
                "SUGGESTIONS_FAILED",
                "ANALYTICS_FAILED",
                "CACHE_CLEAR_FAILED",
                "INVALID_DATE_RANGE",
                "DATE_RANGE_TOO_LARGE",
                "HEALTH_CHECK_FAILED",
                "INTERNAL_ERROR"
              ]
            message:
              type: string
            details:
              type: object
            retry_after_seconds:
              type: integer
            suggestions:
              type: array
              items:
                type: string
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
            processing_time_ms:
              type: integer
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...